ARQ-708 Add support for keeping container alive between Maven execution phases
Bind the individual commands (start/deploy/undeploy/stop) to different Maven phases to control when the container is started and when to deploy. deploy/undeploy is not required if you have other means of deploying to the server. deploy/undeploy will throw Exception if start has not been called.

Two new commands added: deployRemote and undeployRemote
These are indented to be used when dealing with Remote containers only. Start and Stop will be called before and after the command it self. For a remote container this translates to connect/(deploy|undeploy)/disconnect.

Tested by calling deployed servlet between phases